web_interface/requirements.txt
Flask==3.0.0
PyYAML==6.0.2
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
pandas==2.3.3
openpyxl==3.1.5
requests==2.31.0
watchdog==6.0.0
APScheduler==3.11.0
tenacity==9.1.2